Add virtual filesystem for admin module lua files
