+++ cli/admin/common.go
@@ -8,6 +8,7 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/tarantool/cartridge-cli/cli/codegen/static"
 	"github.com/tarantool/cartridge-cli/cli/common"
 	"github.com/tarantool/cartridge-cli/cli/connector"
 	"github.com/tarantool/cartridge-cli/cli/templates"
@@ -49,6 +50,11 @@ func (funcInfo *FuncInfo) Format() string {
 	}
 
 	argsUsageStr := common.FormatStringStringMap(argsUsagesMap)
+
+	funcHelpMsgTmpl, err := static.GetStaticFileContent(AdminLuaTemplateFS, "func_help_msg.lua")
+	if err != nil {
+		panic(err)
+	}
 
 	funcHelpMsg, err := templates.GetTemplatedStr(&funcHelpMsgTmpl, map[string]interface{}{
 		"FuncInfo":  funcInfo.Usage,
@@ -206,6 +212,11 @@ func normalizeFlagName(name string) string {
 }
 
 func getAdminFuncEvalTypedBody(adminFuncName string) (string, error) {
+	evalFuncGetResBodyTmpl, err := static.GetStaticFileContent(AdminLuaTemplateFS, "eval_func_get_res_body.lua")
+	if err != nil {
+		return "", project.InternalError("Failed to parse template: %s", err)
+	}
+
 	funcBody, err := templates.GetTemplatedStr(&evalFuncGetResBodyTmpl, map[string]string{
 		"FuncName": adminFuncName,
 	})
@@ -224,11 +235,3 @@ func getCliExtError(format string, a ...interface{}) error {
 	msg := fmt.Sprintf(format, a...)
 	return fmt.Errorf(cliExtErrFmt, msg)
 }
-
-var (
-	evalFuncGetResBodyTmpl = `
-local res, err = {{ .FuncName }}(...)
-assert(err == nil, err)
-return res
-`
-)
