+++ cli/admin/templates/admin_list_func_body.lua
@@ -1,0 +1,4 @@
+
+local func_list, err = {{ .AdminListFuncName }}(...)
+assert(err == nil, err)
+return func_help
