+++ cli/admin/templates/eval_func_get_res_body.lua
@@ -1,0 +1,4 @@
+
+local res, err = {{ .FuncName }}(...)
+assert(err == nil, err)
+return res
