+++ cli/admin/templates/func_help_msg.lua
@@ -1,0 +1,4 @@
+{{ .FuncInfo }}{{ if .ArgsUsage }}
+
+Args:
+{{ .ArgsUsage }}{{ end }}
