+++ cli/codegen/static/static.go
@@ -1,0 +1,13 @@
+// +build dev
+
+package static
+
+//go:generate go run -tags=dev ../generate_vfs.go -o templatexxx.go
+
+import (
+	"net/http"
+	"path/filepath"
+)
+
+var CartridgeTemplateFS http.FileSystem = http.Dir(filepath.Join("..", "..", "create", "templates", "cartridge"))
+var AdminLuaTemplateFS http.FileSystem = http.Dir(filepath.Join("..", "..", "admin", "templates"))
