--- cli/create/codegen/static/static.go
@@ -1,12 +1,0 @@
-// +build dev
-
-package static
-
-//go:generate go run -tags=dev ../generate_vfs.go -o templatexxx.go
-
-import (
-	"net/http"
-	"path/filepath"
-)
-
-var CartridgeTemplateFS http.FileSystem = http.Dir(filepath.Join("..", "..", "templates", "cartridge"))
